requirements.txt
streamlit
yfinance
groq
pandas
numpy
matplotlib
python-dotenv
scipy
